requirements.txt
accelerate==1.1.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
Django==5.0
django-tinymce==4.1.0
docker-pycreds==0.4.0
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
language_tool_python==2.8.1
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.6
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyMySQL==1.1.1
pypdf==5.1.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sqlparse==0.5.2
svglib==1.5.1
sympy==1.13.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
wandb==0.18.7
webencodings==0.5.1
xhtml2pdf==0.2.16
